final update with minor changes

+++ src/app/pages/layout/layout.page.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { headerConfigKeys } from 'src/app/core/constants';
 import { menuLabelKeys } from 'src/app/core/constants/menu.keys';
 import { LocalStorageService } from 'src/app/core/services';
+import { AlertService } from 'src/app/core/services/alert/alert.service';
 import { UtilService } from 'src/app/shared/util.service';
 
 @Component({
@@ -22,7 +23,8 @@ export class LayoutPage implements OnInit {
   constructor(
     private utlService: UtilService,
     private router: Router,
-    private localStorageService: LocalStorageService
+    private localStorageService: LocalStorageService,
+    private alert: AlertService
   ) {
     this.utlService.setHeaders(this.configHeader);
     this.utlService.configHeader.subscribe((header: any)=>{
@@ -59,11 +61,18 @@ export class LayoutPage implements OnInit {
 
   configMenu = menuLabelKeys;
 
-  handleMenuClicks(action: string){
+  async handleMenuClicks(action: string){
     switch(action){
       case 'SIDE_MENU_LAYOUT.LOGOUT' :
-        this.localStorageService.deleteAll();
-        this.router.navigate(['/auth/login'], { replaceUrl: true });
+        const result = await this.alert.presentAlert('Do you want to log out?');
+        if (result) {
+          // User clicked "Yes"
+          this.localStorageService.deleteAll();
+          this.router.navigate(['/auth/login'], { replaceUrl: true });
+        } else {
+          // User clicked "No"
+          console.log('no');
+        }
         break;
       case 'SIDE_MENU_LAYOUT.HOME':
         this.router.navigate(['/layout/home'], { replaceUrl: true });
